Bug 27263: Link to preferences mentioned in system preference descriptions
This patch attempts to locate all references to system preference names
in various system preference descriptions and link those preference
names to a search for that preference. This will allow the user to
easily jump between preferences which are related or depend on each
other.

To test, apply the patch go to Administration -> System preferences.

- Open each preference category and confirm that the page loads without
  error.
- Test any system preference links you find to confirm that they work
  correctly.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>